Added Refactoring and Unit-Tests
Unit-Tests for the analyzer
